Removed unsed file 'callAPI.js', grouped files into folders
