+++ src/utils/chatHelpers/addRemoveHandler.js
