+++ src/utils/plotHelpers.js/math.js
@@ -1,5 +1,4 @@
 function transpose(matrix) {
     return matrix[0].map((col, c) => matrix.map((row, r) => matrix[r][c]));
 }
-
 export default transpose
